[passthrough-representable] - Deprecated and renamed thenPresent methods with thenProceedWith methods so that the API surface is easier to keep track of - TT
