+++ Sources/SwiftCurrent_UIKit/Extensions/WorkflowUIKitAdditions.swift
@@ -11,65 +11,6 @@
 import Foundation
 import SwiftCurrent
 
-// UI friendly terms for creating a workflow.
-extension Workflow {
-    /**
-     Creates a `Workflow` with a `FlowRepresentable`.
-     - Parameter type: a reference to the first `FlowRepresentable`'s concrete type in the workflow.
-     - Parameter presentationType: the `LaunchStyle.PresentationType` the flow representable should use while it's part of this workflow.
-     - Parameter flowPersistence: the `FlowPersistence` representing how this item in the workflow should persist.
-     */
-    public convenience init(_ type: F.Type,
-                            presentationType: LaunchStyle.PresentationType,
-                            flowPersistence: @escaping @autoclosure () -> FlowPersistence = .default) {
-        self.init(FlowRepresentableMetadata(type,
-                                            launchStyle: presentationType.rawValue) { _ in flowPersistence() })
-    }
-
-    /**
-     Creates a `Workflow` with a `FlowRepresentable`.
-     - Parameter type: a reference to the first `FlowRepresentable`'s concrete type in the workflow.
-     - Parameter presentationType: the `LaunchStyle.PresentationType` the flow representable should use while it's part of this workflow.
-     - Parameter flowPersistence: a closure taking in the `FlowRepresentable.WorkflowInput` and returning a `FlowPersistence` representing how this item in the workflow should persist.
-     */
-    public convenience init(_ type: F.Type,
-                            presentationType: LaunchStyle.PresentationType,
-                            flowPersistence: @escaping (F.WorkflowInput) -> FlowPersistence) {
-        self.init(FlowRepresentableMetadata(type,
-                                            launchStyle: presentationType.rawValue) { data in
-            guard case.args(let extracted) = data,
-                  let cast = extracted as? F.WorkflowInput else { return .default }
-            return flowPersistence(cast)
-        })
-    }
-
-    /**
-     Creates a `Workflow` with a `FlowRepresentable`.
-     - Parameter type: a reference to the first `FlowRepresentable`'s concrete type in the workflow.
-     - Parameter presentationType: the `LaunchStyle.PresentationType` the flow representable should use while it's part of this workflow.
-     - Parameter flowPersistence: a closure returning a `FlowPersistence` representing how this item in the workflow should persist.
-     */
-    public convenience init(_ type: F.Type,
-                            presentationType: LaunchStyle.PresentationType,
-                            flowPersistence: @autoclosure @escaping () -> FlowPersistence) where F.WorkflowInput == Never {
-        self.init(FlowRepresentableMetadata(type,
-                                            launchStyle: presentationType.rawValue) { _ in flowPersistence() })
-    }
-
-    /**
-     Creates a `Workflow` with a `FlowRepresentable`.
-     - Parameter type: a reference to the first `FlowRepresentable`'s concrete type in the workflow.
-     - Parameter presentationType: the `LaunchStyle.PresentationType` the flow representable should use while it's part of this workflow.
-     - Parameter flowPersistence: a closure returning a `FlowPersistence` representing how this item in the workflow should persist.
-     */
-    public convenience init(_ type: F.Type,
-                            presentationType: LaunchStyle.PresentationType,
-                            flowPersistence: @autoclosure @escaping () -> FlowPersistence) where F.WorkflowInput == AnyWorkflow.PassedArgs {
-        self.init(FlowRepresentableMetadata(type,
-                                            launchStyle: presentationType.rawValue) { _ in flowPersistence() })
-    }
-}
-
 extension Workflow {
     /**
      Called when the workflow should be terminated, and the app should return to the point before the workflow was launched.
@@ -106,37 +47,203 @@ extension AnyWorkflow {
     }
 }
 
+extension Workflow {
+    /**
+     Creates a `Workflow` with a `FlowRepresentable`.
+     - Parameter type: a reference to the first `FlowRepresentable`'s concrete type in the workflow.
+     - Parameter launchStyle: the `LaunchStyle` the `FlowRepresentable` should use while it's part of this workflow.
+     - Parameter flowPersistence: a `FlowPersistence` representing how this item in the workflow should persist.
+     */
+    public convenience init(_ type: F.Type,
+                            launchStyle: LaunchStyle.PresentationType,
+                            flowPersistence: @escaping @autoclosure () -> FlowPersistence = .default) {
+        self.init(FlowRepresentableMetadata(type,
+                                            launchStyle: launchStyle.rawValue) { _ in flowPersistence() })
+    }
+
+    /**
+     Creates a `Workflow` with a `FlowRepresentable`.
+     - Parameter type: a reference to the first `FlowRepresentable`'s concrete type in the workflow.
+     - Parameter launchStyle: the `LaunchStyle` the `FlowRepresentable` should use while it's part of this workflow.
+     - Parameter flowPersistence: a closure taking in the generic type from the `FlowRepresentable` and returning a `FlowPersistence` representing how this item in the workflow should persist.
+     */
+    public convenience init(_ type: F.Type,
+                            launchStyle: LaunchStyle.PresentationType,
+                            flowPersistence: @escaping (F.WorkflowInput) -> FlowPersistence) {
+        self.init(FlowRepresentableMetadata(type,
+                                            launchStyle: launchStyle.rawValue) { data in
+            guard case .args(let extracted) = data,
+                  let cast = extracted as? F.WorkflowInput else { return .default }
+            return flowPersistence(cast)
+        })
+    }
+
+    /**
+     Creates a `Workflow` with a `FlowRepresentable`.
+     - Parameter type: a reference to the first `FlowRepresentable`'s concrete type in the workflow.
+     - Parameter launchStyle: the `LaunchStyle` the `FlowRepresentable` should use while it's part of this workflow.
+     - Parameter flowPersistence: a closure returning a `FlowPersistence` representing how this item in the workflow should persist.
+     */
+    public convenience init(_ type: F.Type,
+                            launchStyle: LaunchStyle.PresentationType,
+                            flowPersistence: @escaping () -> FlowPersistence) where F.WorkflowInput == Never {
+        self.init(FlowRepresentableMetadata(type,
+                                            launchStyle: launchStyle.rawValue) { _ in flowPersistence() })
+    }
+
+    /**
+     Creates a `Workflow` with a `FlowRepresentable`.
+     - Parameter type: a reference to the first `FlowRepresentable`'s concrete type in the workflow.
+     - Parameter launchStyle: the `LaunchStyle` the `FlowRepresentable` should use while it's part of this workflow.
+     - Parameter flowPersistence: a closure returning a `FlowPersistence` representing how this item in the workflow should persist.
+     */
+    public convenience init(_ type: F.Type,
+                            launchStyle: LaunchStyle.PresentationType,
+                            flowPersistence: @escaping (F.WorkflowInput) -> FlowPersistence) where F.WorkflowInput == AnyWorkflow.PassedArgs {
+        self.init(FlowRepresentableMetadata(type,
+                                            launchStyle: launchStyle.rawValue) { flowPersistence($0) })
+    }
+}
+
 extension Workflow where F.WorkflowOutput == Never {
     /**
      Adds an item to the workflow; enforces the `FlowRepresentable.WorkflowOutput` of the previous item matches the `FlowRepresentable.WorkflowInput` of this item.
      - Parameter type: a reference to the next `FlowRepresentable`'s concrete type in the workflow.
-     - Parameter presentationType: the `LaunchStyle.PresentationType` the `FlowRepresentable` should use while it's part of this workflow.
+     - Parameter launchStyle: the `LaunchStyle` the `FlowRepresentable` should use while it's part of this workflow.
      - Parameter flowPersistence: a `FlowPersistence` representing how this item in the workflow should persist.
      - Returns: a new workflow with the additional `FlowRepresentable` item.
      */
-    public func thenPresent<FR: FlowRepresentable>(_ type: FR.Type,
-                                                   presentationType: LaunchStyle.PresentationType = .default,
+    public func thenProceed<FR: FlowRepresentable>(with type: FR.Type,
+                                                   launchStyle: LaunchStyle.PresentationType,
                                                    flowPersistence: @escaping @autoclosure () -> FlowPersistence = .default) -> Workflow<FR> where FR.WorkflowInput == Never {
+        thenProceed(with: type,
+                    launchStyle: launchStyle.rawValue,
+                    flowPersistence: flowPersistence)
+    }
+
+    /**
+     Adds an item to the workflow; enforces the `FlowRepresentable.WorkflowOutput` of the previous item matches the `FlowRepresentable.WorkflowInput` of this item.
+     - Parameter type: a reference to the next `FlowRepresentable`'s concrete type in the workflow.
+     - Parameter launchStyle: the `LaunchStyle` the `FlowRepresentable` should use while it's part of this workflow.
+     - Parameter flowPersistence: a `FlowPersistence` representing how this item in the workflow should persist.
+     - Returns: a new workflow with the additional `FlowRepresentable` item.
+     */
+    public func thenProceed<FR: FlowRepresentable>(with type: FR.Type,
+                                                   launchStyle: LaunchStyle.PresentationType,
+                                                   flowPersistence: @escaping () -> FlowPersistence) -> Workflow<FR> where FR.WorkflowInput == Never {
+        thenProceed(with: type,
+                    launchStyle: launchStyle.rawValue,
+                    flowPersistence: flowPersistence)
+    }
+
+    /**
+     Adds an item to the workflow; enforces the `FlowRepresentable.WorkflowOutput` of the previous item matches the `FlowRepresentable.WorkflowInput` of this item.
+     - Parameter type: a reference to the next `FlowRepresentable`'s concrete type in the workflow.
+     - Parameter launchStyle: the `LaunchStyle` the `FlowRepresentable` should use while it's part of this workflow.
+     - Parameter flowPersistence: a `FlowPersistence` representing how this item in the workflow should persist.
+     - Returns: a new workflow with the additional `FlowRepresentable` item.
+     */
+    public func thenProceed<FR: FlowRepresentable>(with type: FR.Type,
+                                                   launchStyle: LaunchStyle.PresentationType,
+                                                   flowPersistence: @escaping @autoclosure () -> FlowPersistence = .default) -> Workflow<FR> where FR.WorkflowInput == AnyWorkflow.PassedArgs {
         let wf = Workflow<FR>(first)
         wf.append(FlowRepresentableMetadata(type,
-                                            launchStyle: presentationType.rawValue) { _ in flowPersistence() })
+                                            launchStyle: launchStyle.rawValue) { _ in flowPersistence() })
+        return wf
+    }
+}
+
+extension Workflow where F.WorkflowOutput == AnyWorkflow.PassedArgs {
+    /**
+     Adds an item to the workflow; enforces the `FlowRepresentable.WorkflowOutput` of the previous item matches the `FlowRepresentable.WorkflowInput` of this item.
+     - Parameter type: a reference to the next `FlowRepresentable`'s concrete type in the workflow.
+     - Parameter launchStyle: the `LaunchStyle` the `FlowRepresentable` should use while it's part of this workflow.
+     - Parameter flowPersistence: a `FlowPersistence` representing how this item in the workflow should persist.
+     - Returns: a new workflow with the additional `FlowRepresentable` item.
+     */
+    public func thenProceed<FR: FlowRepresentable>(with type: FR.Type,
+                                                   launchStyle: LaunchStyle.PresentationType,
+                                                   flowPersistence: @escaping @autoclosure () -> FlowPersistence = .default) -> Workflow<FR> {
+        let wf = Workflow<FR>(first)
+        wf.append(FlowRepresentableMetadata(type,
+                                            launchStyle: launchStyle.rawValue) { _ in flowPersistence() })
         return wf
     }
 
     /**
      Adds an item to the workflow; enforces the `FlowRepresentable.WorkflowOutput` of the previous item matches the `FlowRepresentable.WorkflowInput` of this item.
      - Parameter type: a reference to the next `FlowRepresentable`'s concrete type in the workflow.
-     - Parameter presentationType: the `LaunchStyle.PresentationType` the `FlowRepresentable` should use while it's part of this workflow.
-     - Parameter flowPersistence: a `FlowPersistence` representing how this item in the workflow should persist.
+     - Parameter launchStyle: the `LaunchStyle` the `FlowRepresentable` should use while it's part of this workflow.
+     - Parameter flowPersistence: a closure taking in the generic type from the `FlowRepresentable.WorkflowInput` and returning a `FlowPersistence` representing how this item in the workflow should persist.
      - Returns: a new workflow with the additional `FlowRepresentable` item.
      */
-    public func thenPresent<FR: FlowRepresentable>(_ type: FR.Type,
-                                                   presentationType: LaunchStyle.PresentationType = .default,
+    public func thenProceed<FR: FlowRepresentable>(with type: FR.Type,
+                                                   launchStyle: LaunchStyle.PresentationType,
+                                                   flowPersistence: @escaping (FR.WorkflowInput) -> FlowPersistence) -> Workflow<FR> {
+        thenProceed(with: type,
+                    launchStyle: launchStyle.rawValue,
+                    flowPersistence: flowPersistence)
+    }
+
+    /**
+     Adds an item to the workflow; enforces the `FlowRepresentable.WorkflowOutput` of the previous item matches the `FlowRepresentable.WorkflowInput` of this item.
+     - Parameter type: a reference to the next `FlowRepresentable`'s concrete type in the workflow.
+     - Parameter launchStyle: the `LaunchStyle` the `FlowRepresentable` should use while it's part of this workflow.
+     - Parameter flowPersistence: a closure taking in the generic type from the `FlowRepresentable.WorkflowInput` and returning a `FlowPersistence` representing how this item in the workflow should persist.
+     - Returns: a new workflow with the additional `FlowRepresentable` item.
+     */
+    public func thenProceed<FR: FlowRepresentable>(with type: FR.Type,
+                                                   launchStyle: LaunchStyle.PresentationType,
+                                                   flowPersistence: @escaping (FR.WorkflowInput) -> FlowPersistence) -> Workflow<FR> where FR.WorkflowInput == AnyWorkflow.PassedArgs {
+        thenProceed(with: type,
+                    launchStyle: launchStyle.rawValue,
+                    flowPersistence: flowPersistence)
+    }
+
+    /**
+     Adds an item to the workflow; enforces the `FlowRepresentable.WorkflowOutput` of the previous item matches the `FlowRepresentable.WorkflowInput` of this item.
+     - Parameter type: a reference to the next `FlowRepresentable`'s concrete type in the workflow.
+     - Parameter launchStyle: the `LaunchStyle` the `FlowRepresentable` should use while it's part of this workflow.
+     - Parameter flowPersistence: a closure taking in the generic type from the `FlowRepresentable.WorkflowInput` and returning a `FlowPersistence` representing how this item in the workflow should persist.
+     - Returns: a new workflow with the additional `FlowRepresentable` item.
+     */
+    public func thenProceed<FR: FlowRepresentable>(with type: FR.Type,
+                                                   launchStyle: LaunchStyle.PresentationType,
                                                    flowPersistence: @escaping @autoclosure () -> FlowPersistence = .default) -> Workflow<FR> where FR.WorkflowInput == AnyWorkflow.PassedArgs {
         let wf = Workflow<FR>(first)
         wf.append(FlowRepresentableMetadata(type,
-                                            launchStyle: presentationType.rawValue) { _ in flowPersistence() })
+                                            launchStyle: launchStyle.rawValue) { _ in flowPersistence() })
         return wf
+    }
+
+    /**
+     Adds an item to the workflow; enforces the `FlowRepresentable.WorkflowOutput` of the previous item matches the `FlowRepresentable.WorkflowInput` of this item.
+     - Parameter type: a reference to the next `FlowRepresentable`'s concrete type in the workflow.
+     - Parameter launchStyle: the `LaunchStyle` the `FlowRepresentable` should use while it's part of this workflow.
+     - Parameter flowPersistence: a closure returning a `FlowPersistence` representing how this item in the workflow should persist.
+     - Returns: a new workflow with the additional `FlowRepresentable` item.
+     */
+    public func thenProceed<FR: FlowRepresentable>(with type: FR.Type,
+                                                   launchStyle: LaunchStyle.PresentationType,
+                                                   flowPersistence: @escaping @autoclosure () -> FlowPersistence = .default) -> Workflow<FR> where FR.WorkflowInput == Never {
+        thenProceed(with: type,
+                    launchStyle: launchStyle.rawValue,
+                    flowPersistence: flowPersistence)
+    }
+
+    /**
+     Adds an item to the workflow; enforces the `FlowRepresentable.WorkflowOutput` of the previous item matches the `FlowRepresentable.WorkflowInput` of this item.
+     - Parameter type: a reference to the next `FlowRepresentable`'s concrete type in the workflow.
+     - Parameter launchStyle: the `LaunchStyle` the `FlowRepresentable` should use while it's part of this workflow.
+     - Parameter flowPersistence: a closure returning a `FlowPersistence` representing how this item in the workflow should persist.
+     - Returns: a new workflow with the additional `FlowRepresentable` item.
+     */
+    public func thenProceed<FR: FlowRepresentable>(with type: FR.Type,
+                                                   launchStyle: LaunchStyle.PresentationType,
+                                                   flowPersistence: @escaping () -> FlowPersistence) -> Workflow<FR> where FR.WorkflowInput == Never {
+        thenProceed(with: type,
+                    launchStyle: launchStyle.rawValue,
+                    flowPersistence: flowPersistence)
     }
 }
 
@@ -144,68 +251,78 @@ extension Workflow {
     /**
      Adds an item to the workflow; enforces the `FlowRepresentable.WorkflowOutput` of the previous item matches the `FlowRepresentable.WorkflowInput` of this item.
      - Parameter type: a reference to the next `FlowRepresentable`'s concrete type in the workflow.
-     - Parameter presentationType: the `LaunchStyle.PresentationType` the `FlowRepresentable` should use while it's part of this workflow.
+     - Parameter launchStyle: the `LaunchStyle` the `FlowRepresentable` should use while it's part of this workflow.
      - Parameter flowPersistence: a `FlowPersistence` representing how this item in the workflow should persist.
      - Returns: a new workflow with the additional `FlowRepresentable` item.
      */
-    public func thenPresent<FR: FlowRepresentable>(_ type: FR.Type,
-                                                   presentationType: LaunchStyle.PresentationType = .default,
+    public func thenProceed<FR: FlowRepresentable>(with type: FR.Type,
+                                                   launchStyle: LaunchStyle.PresentationType,
                                                    flowPersistence: @escaping @autoclosure () -> FlowPersistence = .default) -> Workflow<FR> where F.WorkflowOutput == FR.WorkflowInput {
         let wf = Workflow<FR>(first)
         wf.append(FlowRepresentableMetadata(type,
-                                            launchStyle: presentationType.rawValue) { _ in flowPersistence() })
+                                            launchStyle: launchStyle.rawValue) { _ in flowPersistence() })
         return wf
     }
 
     /**
      Adds an item to the workflow; enforces the `FlowRepresentable.WorkflowOutput` of the previous item matches the `FlowRepresentable.WorkflowInput` of this item.
      - Parameter type: a reference to the next `FlowRepresentable`'s concrete type in the workflow.
-     - Parameter presentationType: the `LaunchStyle.PresentationType` the `FlowRepresentable` should use while it's part of this workflow.
-     - Parameter flowPersistence: a closure taking in the `FlowRepresentable.WorkflowInput` and returning a `FlowPersistence` representing how this item in the workflow should persist.
+     - Parameter launchStyle: the `LaunchStyle` the `FlowRepresentable` should use while it's part of this workflow.
+     - Parameter flowPersistence: a closure taking in the generic type from the `FlowRepresentable.WorkflowInput` and returning a `FlowPersistence` representing how this item in the workflow should persist.
      - Returns: a new workflow with the additional `FlowRepresentable` item.
      */
-    public func thenPresent<FR: FlowRepresentable>(_ type: FR.Type,
-                                                   presentationType: LaunchStyle.PresentationType = .default,
+    public func thenProceed<FR: FlowRepresentable>(with type: FR.Type,
+                                                   launchStyle: LaunchStyle.PresentationType,
                                                    flowPersistence: @escaping (FR.WorkflowInput) -> FlowPersistence) -> Workflow<FR> where F.WorkflowOutput == FR.WorkflowInput {
-        let wf = Workflow<FR>(first)
-        wf.append(FlowRepresentableMetadata(type,
-                                            launchStyle: presentationType.rawValue) { data in
-            guard case.args(let extracted) = data,
-                  let cast = extracted as? FR.WorkflowInput else { return .default }
-            return flowPersistence(cast)
-        })
-        return wf
+        thenProceed(with: type,
+                    launchStyle: launchStyle.rawValue,
+                    flowPersistence: flowPersistence)
     }
 
     /**
      Adds an item to the workflow; enforces the `FlowRepresentable.WorkflowOutput` of the previous item matches the `FlowRepresentable.WorkflowInput` of this item.
      - Parameter type: a reference to the next `FlowRepresentable`'s concrete type in the workflow.
-     - Parameter presentationType: the `LaunchStyle.PresentationType` the `FlowRepresentable` should use while it's part of this workflow.
-     - Parameter flowPersistence: a `FlowPersistence` representing how this item in the workflow should persist.
+     - Parameter launchStyle: the `LaunchStyle` the `FlowRepresentable` should use while it's part of this workflow.
+     - Parameter flowPersistence: a closure returning a `FlowPersistence` representing how this item in the workflow should persist.
      - Returns: a new workflow with the additional `FlowRepresentable` item.
      */
-    public func thenPresent<FR: FlowRepresentable>(_ type: FR.Type,
-                                                   presentationType: LaunchStyle.PresentationType = .default,
+    public func thenProceed<FR: FlowRepresentable>(with type: FR.Type,
+                                                   launchStyle: LaunchStyle.PresentationType,
                                                    flowPersistence: @escaping @autoclosure () -> FlowPersistence = .default) -> Workflow<FR> where FR.WorkflowInput == Never {
         let wf = Workflow<FR>(first)
         wf.append(FlowRepresentableMetadata(type,
-                                            launchStyle: presentationType.rawValue) { _ in flowPersistence() })
+                                            launchStyle: launchStyle.rawValue) { _ in flowPersistence() })
         return wf
     }
 
     /**
      Adds an item to the workflow; enforces the `FlowRepresentable.WorkflowOutput` of the previous item matches the `FlowRepresentable.WorkflowInput` of this item.
      - Parameter type: a reference to the next `FlowRepresentable`'s concrete type in the workflow.
-     - Parameter presentationType: the `LaunchStyle.PresentationType` the `FlowRepresentable` should use while it's part of this workflow.
-     - Parameter flowPersistence: a `FlowPersistence` representing how this item in the workflow should persist.
+     - Parameter launchStyle: the `LaunchStyle` the `FlowRepresentable` should use while it's part of this workflow.
+     - Parameter flowPersistence: a closure returning a `FlowPersistence` representing how this item in the workflow should persist.
      - Returns: a new workflow with the additional `FlowRepresentable` item.
      */
-    public func thenPresent<FR: FlowRepresentable>(_ type: FR.Type,
-                                                   presentationType: LaunchStyle.PresentationType = .default,
+    public func thenProceed<FR: FlowRepresentable>(with type: FR.Type,
+                                                   launchStyle: LaunchStyle.PresentationType,
                                                    flowPersistence: @escaping @autoclosure () -> FlowPersistence = .default) -> Workflow<FR> where FR.WorkflowInput == AnyWorkflow.PassedArgs {
         let wf = Workflow<FR>(first)
         wf.append(FlowRepresentableMetadata(type,
-                                            launchStyle: presentationType.rawValue) { _ in flowPersistence() })
+                                            launchStyle: launchStyle.rawValue) { _ in flowPersistence() })
         return wf
+    }
+
+    /**
+     Adds an item to the workflow; enforces the `FlowRepresentable.WorkflowOutput` of the previous item matches the `FlowRepresentable.WorkflowInput` of this item.
+     - Parameter type: a reference to the next `FlowRepresentable`'s concrete type in the workflow.
+     - Parameter launchStyle: the `LaunchStyle` the `FlowRepresentable` should use while it's part of this workflow.
+     - Parameter flowPersistence: a closure returning a `FlowPersistence` representing how this item in the workflow should persist.
+     - Returns: a new workflow with the additional `FlowRepresentable` item.
+     */
+    public func thenProceed<FR: FlowRepresentable>(with type: FR.Type,
+                                                   launchStyle: LaunchStyle.PresentationType,
+                                                   flowPersistence: @escaping (FR.WorkflowInput) -> FlowPersistence) -> Workflow<FR> where FR.WorkflowInput == AnyWorkflow.PassedArgs {
+        thenProceed(with: type,
+                    launchStyle: launchStyle.rawValue,
+                    flowPersistence: flowPersistence)
     }
 }
