+++ Tests/SwiftCurrentTests/FlowPersistenceTests.swift
@@ -1,8 +1,9 @@
 //
 //  FlowPersistenceTests.swift
-//  
+//  SwiftCurrentTests
 //
 //  Created by Tyler Thompson on 11/26/20.
+//  Copyright © 2021 WWT and Tyler Thompson. All rights reserved.
 //
 
 import Foundation
